Extract new unshrinkable class
